refactor common methods into sql_utils
